+++ src/components/compositions/Record.vue
@@ -14,6 +14,7 @@ import type {
   Dates as IsoDates,
   DistributionOption as IsoDistributionOption,
   Extent as IsoExtent,
+  GraphicOverview as IsoGraphicOverview,
   Identifier,
   Lineage as IsoLineage,
   PointOfContact as IsoContact,
@@ -38,6 +39,7 @@ import Lineage from '@/components/sections/elements/Lineage.vue'
 import ResourceType from '@/components/sections/elements/ResourceType.vue'
 import Services from '@/components/sections/elements/Services.vue'
 import Summary from '@/components/sections/elements/Summary.vue'
+import Thumbnails from '@/components/sections/elements/Thumbnails.vue'
 import Title from '@/components/sections/elements/Title.vue'
 
 defineProps({
@@ -60,6 +62,8 @@ const magicPoC = createOrgSlugPointOfContact('bas_magic', 'pointOfContact')
 let contacts: ComputedRef<IsoContact[]> = computed(() => {
   return [...authors.value, magicPoC]
 })
+
+const thumbnails = ref<IsoGraphicOverview[]>([])
 
 const accessConstraint = ref<IsoConstraint | null>(null)
 const licenceConstraint = ref<IsoConstraint | null>(null)
@@ -104,6 +108,17 @@ let isoRecordMerged: ComputedRef<IsoRecord> = computed(() => {
       identification: {
         ...mergedRecord.identification,
         contacts: contacts.value,
+      },
+    }
+  }
+
+  // merge isoRecord and thumbnails at 'isoRecord.identification.graphic_overviews' if not empty
+  if (thumbnails.value.length > 0) {
+    mergedRecord = {
+      ...mergedRecord,
+      identification: {
+        ...mergedRecord.identification,
+        graphic_overviews: thumbnails.value,
       },
     }
   }
@@ -217,6 +232,10 @@ watch(
       @update:iso-other-citation-details="
         (event: string) => (isoRecord.identification.other_citation_details = event)
       "
+    />
+    <Thumbnails
+      :file-identifier="record.fileIdentifier"
+      @update:iso-graphic-overviews="(event: IsoGraphicOverview[]) => (thumbnails = event)"
     />
     <Downloads
       v-if="show('downloads')"